move decryptionResults to a separate file
