+++ .gitignore
@@ -1,6 +1,5 @@
 */node_modules/*
 precompiles/cache/*
-precompiles/types/*
 precompiles/artifacts/*
 solgen/**/*.js
 build/
